+++ src/LightsprintCore/RRCollider/IntersectBspFast.cpp
@@ -782,7 +782,7 @@ bool IntersectBspFast IBP2::isValidTriangle(unsigned i) const
 }
 
 template IBP
-unsigned IntersectBspFast IBP2::getMemoryOccupied() const
+size_t IntersectBspFast IBP2::getMemoryOccupied() const
 {
 	return sizeof(IntersectBspFast) 
 		+ (tree?tree->bsp.size:0)
